Crate google_mybusiness4

This documentation was generated from My Business crate version 2.0.0+0, where 0 is the exact revision of the mybusiness:v4 schema built by the mako code generator v2.0.0.

Everything else about the My Business v4 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

accounts

admins create, admins delete, admins list, admins patch, create, delete notifications, generate account number, get, get notifications, invitations accept, invitations decline, invitations list, list, list recommend google locations, locations admins create, locations admins delete, locations admins list, locations admins patch, locations associate, locations batch get, locations batch get reviews, locations clear association, locations create, locations delete, locations fetch verification options, locations find matches, locations followers get metadata, locations get, locations get google updated, locations list, locations local posts create, locations local posts delete, locations local posts get, locations local posts list, locations local posts patch, locations local posts report insights, locations media create, locations media customers get, locations media customers list, locations media delete, locations media get, locations media list, locations media patch, locations media start upload, locations patch, locations questions answers delete, locations questions answers list, locations questions answers upsert, locations questions create, locations questions delete, locations questions list, locations questions patch, locations report insights, locations reviews delete reply, locations reviews get, locations reviews list, locations reviews update reply, locations transfer, locations verifications complete, locations verifications list, locations verify, update and update notifications

attributes

list

categories

list

chains

get and search

google locations

report and search

verification tokens

generate

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit().await

Or specifically ...

let r = hub.accounts().admins_create(...).doit().await
let r = hub.accounts().admins_delete(...).doit().await
let r = hub.accounts().admins_list(...).doit().await
let r = hub.accounts().admins_patch(...).doit().await
let r = hub.accounts().invitations_accept(...).doit().await
let r = hub.accounts().invitations_decline(...).doit().await
let r = hub.accounts().invitations_list(...).doit().await
let r = hub.accounts().locations_admins_create(...).doit().await
let r = hub.accounts().locations_admins_delete(...).doit().await
let r = hub.accounts().locations_admins_list(...).doit().await
let r = hub.accounts().locations_admins_patch(...).doit().await
let r = hub.accounts().locations_followers_get_metadata(...).doit().await
let r = hub.accounts().locations_local_posts_create(...).doit().await
let r = hub.accounts().locations_local_posts_delete(...).doit().await
let r = hub.accounts().locations_local_posts_get(...).doit().await
let r = hub.accounts().locations_local_posts_list(...).doit().await
let r = hub.accounts().locations_local_posts_patch(...).doit().await
let r = hub.accounts().locations_local_posts_report_insights(...).doit().await
let r = hub.accounts().locations_media_customers_get(...).doit().await
let r = hub.accounts().locations_media_customers_list(...).doit().await
let r = hub.accounts().locations_media_create(...).doit().await
let r = hub.accounts().locations_media_delete(...).doit().await
let r = hub.accounts().locations_media_get(...).doit().await
let r = hub.accounts().locations_media_list(...).doit().await
let r = hub.accounts().locations_media_patch(...).doit().await
let r = hub.accounts().locations_media_start_upload(...).doit().await
let r = hub.accounts().locations_questions_answers_delete(...).doit().await
let r = hub.accounts().locations_questions_answers_list(...).doit().await
let r = hub.accounts().locations_questions_answers_upsert(...).doit().await
let r = hub.accounts().locations_questions_create(...).doit().await
let r = hub.accounts().locations_questions_delete(...).doit().await
let r = hub.accounts().locations_questions_list(...).doit().await
let r = hub.accounts().locations_questions_patch(...).doit().await
let r = hub.accounts().locations_reviews_delete_reply(...).doit().await
let r = hub.accounts().locations_reviews_get(...).doit().await
let r = hub.accounts().locations_reviews_list(...).doit().await
let r = hub.accounts().locations_reviews_update_reply(...).doit().await
let r = hub.accounts().locations_verifications_complete(...).doit().await
let r = hub.accounts().locations_verifications_list(...).doit().await
let r = hub.accounts().locations_associate(...).doit().await
let r = hub.accounts().locations_batch_get(...).doit().await
let r = hub.accounts().locations_batch_get_reviews(...).doit().await
let r = hub.accounts().locations_clear_association(...).doit().await
let r = hub.accounts().locations_create(...).doit().await
let r = hub.accounts().locations_delete(...).doit().await
let r = hub.accounts().locations_fetch_verification_options(...).doit().await
let r = hub.accounts().locations_find_matches(...).doit().await
let r = hub.accounts().locations_get(...).doit().await
let r = hub.accounts().locations_get_google_updated(...).doit().await
let r = hub.accounts().locations_list(...).doit().await
let r = hub.accounts().locations_patch(...).doit().await
let r = hub.accounts().locations_report_insights(...).doit().await
let r = hub.accounts().locations_transfer(...).doit().await
let r = hub.accounts().locations_verify(...).doit().await
let r = hub.accounts().create(...).doit().await
let r = hub.accounts().delete_notifications(...).doit().await
let r = hub.accounts().generate_account_number(...).doit().await
let r = hub.accounts().get(...).doit().await
let r = hub.accounts().get_notifications(...).doit().await
let r = hub.accounts().list(...).doit().await
let r = hub.accounts().list_recommend_google_locations(...).doit().await
let r = hub.accounts().update(...).doit().await
let r = hub.accounts().update_notifications(...).doit().await

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-mybusiness4 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.14"
hyper-rustls = "^0.22"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^5.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_mybusiness4 as mybusiness4;
use mybusiness4::{Result, Error};
use std::default::Default;
use oauth2;
use mybusiness4::MyBusiness;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = MyBusiness::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().locations_questions_list("parent")
             .page_token("sed")
             .page_size(-2)
             .order_by("takimata")
             .filter("amet.")
             .answers_per_question(-20)
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be encodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Re-exports

pub use api::MyBusiness;

pub use client::Result;

pub use client::Error;

pub use client::Delegate;

Modules

api

client