Module google_mybusiness4::api

Structs

AcceptInvitationRequest
Request message for AccessControl.AcceptInvitation.

Account
An account is a container for your business's locations. If you are the only user who manages locations for your business, you can use your personal Google Account. To share management of locations with multiple users, [create a business account] (https://support.google.com/business/answer/6085339?ref_topic=6085325).

AccountAdminCreateCall
Invites the specified user to become an administrator for the specified account. The invitee must accept the invitation in order to be granted access to the account. See AcceptInvitation to programmatically accept an invitation.

AccountAdminDeleteCall
Removes the specified admin from the specified account.

AccountAdminListCall
Lists the admins for the specified account.

AccountAdminPatchCall
Updates the Admin for the specified Account Admin. Only the AdminRole of the Admin can be updated.

AccountCreateCall
Creates an account with the specified name and type under the given parent.

AccountDeleteNotificationCall
Clears the pubsub notification settings for the account.

AccountGenerateAccountNumberCall
Generates an account number for this account. The account number is not provisioned when an account is created. Use this request to create an account number when it is required.

AccountGetCall
Gets the specified account. Returns NOT_FOUND if the account does not exist or if the caller does not have access rights to it.

AccountGetNotificationCall
Returns the pubsub notification settings for the account.

AccountInvitationAcceptCall
Accepts the specified invitation.

AccountInvitationDeclineCall
Declines the specified invitation.

AccountInvitationListCall
Lists pending invitations for the specified account.

AccountListCall
Lists all of the accounts for the authenticated user. This includes all accounts that the user owns, as well as any accounts for which the user has management rights.

AccountListRecommendGoogleLocationCall
List all the GoogleLocations that have been recommended to the specified GMB account. Recommendations are provided for personal accounts and location groups only, requests for all other account types will result in an error. The recommendations for location groups are based on the locations in that group.

AccountLocationAdminCreateCall
Invites the specified user to become an administrator for the specified location. The invitee must accept the invitation in order to be granted access to the location. See AcceptInvitation to programmatically accept an invitation.

AccountLocationAdminDeleteCall
Removes the specified admin as a manager of the specified location.

AccountLocationAdminListCall
Lists all of the admins for the specified location.

AccountLocationAdminPatchCall
Updates the Admin for the specified Location Admin. Only the AdminRole of the Admin can be updated.

AccountLocationAssociateCall
Associates a location to a place ID. Any previous association is overwritten. This operation is only valid if the location is unverified. The association must be valid, that is, it appears in the list of FindMatchingLocations.

AccountLocationBatchGetCall
Gets all of the specified locations in the given account.

AccountLocationBatchGetReviewCall
Returns the paginated list of reviews for all specified locations. This operation is only valid if the specified locations are verified.

AccountLocationClearAssociationCall
Clears an association between a location and its place ID. This operation is only valid if the location is unverified.

AccountLocationCreateCall
Creates a new location owned by the specified account, and returns it.

AccountLocationDeleteCall
Deletes a location.

AccountLocationFetchVerificationOptionCall
Reports all eligible verification options for a location in a specific language.

AccountLocationFindMatcheCall
Finds all of the possible locations that are a match to the specified location. This operation is only valid if the location is unverified.

AccountLocationFollowerGetMetadataCall
Get the followers settings for a location.

AccountLocationGetCall
Gets the specified location. Returns NOT_FOUND if the location does not exist.

AccountLocationGetGoogleUpdatedCall
Gets the Google-updated version of the specified location. Returns NOT_FOUND if the location does not exist.

AccountLocationListCall
Lists the locations for the specified account.

AccountLocationLocalPostCreateCall
Creates a new local post associated with the specified location, and returns it.

AccountLocationLocalPostDeleteCall
Deletes a local post. Returns NOT_FOUND if the local post does not exist.

AccountLocationLocalPostGetCall
Gets the specified local post. Returns NOT_FOUND if the local post does not exist.

AccountLocationLocalPostListCall
Returns a list of local posts associated with a location.

AccountLocationLocalPostPatchCall
Updates the specified local post and returns the updated local post.

AccountLocationLocalPostReportInsightCall
Returns insights for a set of local posts associated with a single listing. Which metrics and how they are reported are options specified in the request proto.

AccountLocationMediaCreateCall
Creates a new media item for the location.

AccountLocationMediaCustomerGetCall
Returns metadata for the requested customer media item.

AccountLocationMediaCustomerListCall
Returns a list of media items associated with a location that have been contributed by customers.

AccountLocationMediaDeleteCall
Deletes the specified media item.

AccountLocationMediaGetCall
Returns metadata for the requested media item.

AccountLocationMediaListCall
Returns a list of media items associated with a location.

AccountLocationMediaPatchCall
Updates metadata of the specified media item. This can only be used to update the Category of a media item, with the exception that the new category cannot be COVER or PROFILE.

AccountLocationMediaStartUploadCall
Generates a MediaItemDataRef for media item uploading.

AccountLocationPatchCall
Updates the specified location.

AccountLocationQuestionAnswerDeleteCall
Deletes the answer written by the current user to a question.

AccountLocationQuestionAnswerListCall
Returns the paginated list of answers for a specified question.

AccountLocationQuestionAnswerUpsertCall
Creates an answer or updates the existing answer written by the user for the specified question. A user can only create one answer per question.

AccountLocationQuestionCreateCall
Adds a question for the specified location.

AccountLocationQuestionDeleteCall
Deletes a specific question written by the current user.

AccountLocationQuestionListCall
Returns the paginated list of questions and some of its answers for a specified location.

AccountLocationQuestionPatchCall
Updates a specific question written by the current user.

AccountLocationReportInsightCall
Returns a report containing insights on one or more metrics by location.

AccountLocationReviewDeleteReplyCall
Deletes the response to the specified review. This operation is only valid if the specified location is verified.

AccountLocationReviewGetCall
Returns the specified review. This operation is only valid if the specified location is verified. Returns NOT_FOUND if the review does not exist, or has been deleted.

AccountLocationReviewListCall
Returns the paginated list of reviews for the specified location. This operation is only valid if the specified location is verified.

AccountLocationReviewUpdateReplyCall
Updates the reply to the specified review. A reply is created if one does not exist. This operation is only valid if the specified location is verified.

AccountLocationTransferCall
Moves a location from an account that the user owns to another account that the same user administers. The user must be an owner of the account the location is currently associated with and must also be at least a manager of the destination account. Returns the Location with its new resource name.

AccountLocationVerificationCompleteCall
Completes a PENDING verification.

AccountLocationVerificationListCall
List verifications of a location, ordered by create time.

AccountLocationVerifyCall
Starts the verification process for a location.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the MyBusiness hub.

AccountState
Indicates status of the account, such as whether the account has been verified by Google.

AccountUpdateCall
Updates the specified business account. Personal accounts cannot be updated using this method.

AccountUpdateNotificationCall
Sets the pubsub notification settings for the account informing My Business which topic to send pubsub notifications for:

AdWordsLocationExtensions
Additional information that is surfaced in AdWords.

AddressInput
Input for ADDRESS verification.

AddressVerificationData
Display data for verifications through postcard.

Admin
An administrator of an Account or a Location.

Answer
Represents an answer to a question

AssociateLocationRequest
Request message for Locations.AssociateLocationRequest.

Attribute
A location attribute. Attributes provide additional information about a location. The attributes that can be set on a location may vary based on the properties of that location (for example, category). Available attributes are determined by Google and may be added and removed without API changes.

AttributeListCall
Returns the list of available attributes that would be available for a location with the given primary category and country.

AttributeMetadata
Metadata for an attribute. Contains display information for the attribute, including a localized name and a heading for grouping related attributes together.

AttributeMethods
A builder providing access to all methods supported on attribute resources. It is not used directly, but through the MyBusiness hub.

AttributeValueMetadata
Metadata for supported attribute values.

Attribution
Attribution information for customer media items, such as the contributor's name and profile picture.

Author
Represents the author of a question or answer

BasicMetricsRequest
A request for basic metric insights.

BatchGetLocationsRequest
Request message for Locations.BatchGetLocations.

BatchGetLocationsResponse
Response message for Locations.BatchGetLocations.

BatchGetReviewsRequest
Request message for Reviews.BatchGetReviews.

BatchGetReviewsResponse
Response message for Reviews.BatchGetReviews.

BusinessHours
Represents the time periods that this location is open for business. Holds a collection of TimePeriod instances.

CallToAction
An action that is performed when the user clicks through the post

Category
A category describing what this business is (not what it does). For a list of valid category IDs, and the mappings to their human-readable names, see categories.list.

CategoryListCall
Returns a list of business categories. Search will match the category name but not the category ID.

CategoryMethods
A builder providing access to all methods supported on category resources. It is not used directly, but through the MyBusiness hub.

Chain
A chain is a brand that your business's locations can be affiliated with.

ChainGetCall
Gets the specified chain. Returns NOT_FOUND if the chain does not exist.

ChainMethods
A builder providing access to all methods supported on chain resources. It is not used directly, but through the MyBusiness hub.

ChainName
Name to be used when displaying the chain.

ChainSearchCall
Searches the chain based on chain name.

ChainUrl
Url to be used when displaying the chain.

ClearLocationAssociationRequest
Request message for Locations.ClearLocationAssociationRequest.

CompleteVerificationRequest
Request message for Verifications.CompleteVerificationAction.

CompleteVerificationResponse
Response message for Verifications.CompleteVerificationAction.

Date
Represents a whole or partial calendar date, e.g. a birthday. The time of day and time zone are either specified elsewhere or are not significant. The date is relative to the Proleptic Gregorian Calendar. This can represent:

DeclineInvitationRequest
Request message for AccessControl.DeclineInvitation.

DimensionalMetricValue
A value for a single metric with a given time dimension.

Dimensions
Dimensions of the media item.

DrivingDirectionMetricsRequest
A request for driving direction insights.

Duplicate
Information about the location that this location duplicates.

EmailInput
Input for EMAIL verification.

EmailVerificationData
Display data for verifications through email.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

FetchVerificationOptionsRequest
Request message for Verifications.FetchVerificationOptions.

FetchVerificationOptionsResponse
Response message for Verifications.FetchVerificationOptions.

FindMatchingLocationsRequest
Request message for Locations.FindMatchingLocations.

FindMatchingLocationsResponse
Response message for Locations.FindMatchingLocations.

FollowersMetadata
Follower metadata for a location.

GenerateAccountNumberRequest
Request message for Accounts.GenerateAccountNumber.

GenerateVerificationTokenRequest
Request message for Verifications.GenerateVerificationToken.

GenerateVerificationTokenResponse
Response message for Verifications.GenerateVerificationToken.

GoogleLocation
Represents a Location that is present on Google. This can be a location that has been claimed by the user, someone else, or could be unclaimed.

GoogleLocationMethods
A builder providing access to all methods supported on googleLocation resources. It is not used directly, but through the MyBusiness hub.

GoogleLocationReportCall
Report a GoogleLocation.

GoogleLocationSearchCall
Search all of the possible locations that are a match to the specified request.

GoogleUpdatedLocation
Represents a location that was modified by Google.

Invitation
Output only. Represents a pending invitation.

Item
A single list item. Each variation of an item in the price list should have its own Item with its own price data.

Label
Label to be used when displaying the price list, section, or item.

LatLng
An object representing a latitude/longitude pair. This is expressed as a pair of doubles representing degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

ListAccountAdminsResponse
Response message for AccessControl.ListAccountAdmins.

ListAccountsResponse
Response message for Accounts.ListAccounts.

ListAnswersResponse
Response message for QuestionsAndAnswers.ListAnswers

ListAttributeMetadataResponse
Response message for Locations.ListAttributeMetadata.

ListBusinessCategoriesResponse
There is no detailed description.

ListCustomerMediaItemsResponse
Response message for Media.ListCustomerMediaItems.

ListInvitationsResponse
Response message for AccessControl.ListInvitations.

ListLocalPostsResponse
Response message for ListLocalPosts

ListLocationAdminsResponse
Response message for AccessControl.ListLocationAdmins.

ListLocationsResponse
Response message for Locations.ListLocations.

ListMediaItemsResponse
Response message for Media.ListMediaItems.

ListQuestionsResponse
Response message for QuestionsAndAnswers.ListQuestions

ListRecommendedGoogleLocationsResponse
Response message for GoogleLocations.ListRecommendedGoogleLocations.

ListReviewsResponse
Response message for Reviews.ListReviews.

ListVerificationsResponse
Response message for Verifications.ListVerifications.

LocalPost
Represents a local post for a location.

LocalPostEvent
All the information pertaining to an event featured in a local post.

LocalPostMetrics
All the metrics requested for a Local Post.

LocalPostOffer
Specific fields for offer posts.

Location
A location. See the [help center article] (https://support.google.com/business/answer/3038177) for a detailed description of these fields, or the category endpoint for a list of valid business categories.

LocationAssociation
How the media item is associated with its location.

LocationDrivingDirectionMetrics
A location indexed with the regions that people usually come from. This is captured by counting how many driving-direction requests to this location are from each region.

LocationKey
Alternate/surrogate key references for a location.

LocationMetrics
A series of Metrics and BreakdownMetrics associated with a Location over some time range.

LocationReview
Represents a review with location information.

LocationState
Contains a set of booleans that reflect the state of a Location.

MatchedLocation
Represents a possible match to a location.

MediaInsights
Insights and statistics for the media item.

MediaItem
A single media item.

MediaItemDataRef
Reference to the photo binary data of a MediaItem uploaded through the My Business API.

Metadata
Additional non-user-editable information about the location.

MetricRequest
A request to return values for one metric and the options for how those values should be returned.

MetricValue
A value for a single Metric from a starting time.

Money
Represents an amount of money with its currency type.

MyBusiness
Central instance to access all MyBusiness related resource activities

Notifications
A Google Cloud Pub/Sub topic where notifications can be published when a location is updated or has a new review. There will be only one notification settings resource per-account.

OpenInfo
Information related to the opening state of the business.

OrganizationInfo
Additional Info stored for an organization.

PhoneInput
Input for PHONE_CALL/SMS verification.

PhoneVerificationData
Display Data for verifications through phone, e.g. phone call, sms.

PlaceInfo
Defines an area that's represented by a place ID.

Places
Defines the union of areas represented by a set of places.

PointRadius
A radius around a particular point (latitude/longitude).

PostalAddress
Represents a postal address, e.g. for postal delivery or payments addresses. Given a postal address, a postal service can deliver items to a premise, P.O. Box or similar. It is not intended to model geographical locations (roads, towns, mountains).

PriceList
A list of item price information. Price lists are structured as one or more price lists, each containing one or more sections with one or more items. For example, food price lists may represent breakfast/lunch/dinner menus, with sections for burgers/steak/seafood.

Profile
All information pertaining to the location's profile.

Question
Represents a single question and some of its answers.

RegionCount
A region with its associated request count.

RelationshipData
Information of all parent and children locations related to this one.

RepeatedEnumAttributeValue
Values for an attribute with a value_type of REPEATED_ENUM. This consists of two lists of value IDs: those that are set (true) and those that are unset (false). Values absent are considered unknown. At least one value must be specified.

ReportGoogleLocationRequest
Request message for reporting a GoogleLocation.

ReportLocalPostInsightsRequest
Request message for Insights.ReportLocalPostInsights

ReportLocalPostInsightsResponse
Response message for Insights.ReportLocalPostInsights

ReportLocationInsightsRequest
Request message for Insights.ReportLocationInsights.

ReportLocationInsightsResponse
Response message for Insights.ReportLocationInsights.

Review
Output only. Represents a review for a location.

ReviewReply
Represents the location owner/manager's reply to a review.

Reviewer
Represents the author of the review.

SearchChainsResponse
Response message for Locations.SearchChains.

SearchGoogleLocationsRequest
Request message for GoogleLocations.SearchGoogleLocations.

SearchGoogleLocationsResponse
Response message for GoogleLocations.SearchGoogleLocations.

Section
A section of the price list containing one or more items.

ServiceAreaBusiness
Service area businesses provide their service at the customer's location (for example, a locksmith or plumber).

ServiceBusinessContext
Additional data for service business verification.

SpecialHourPeriod
Represents a single time period when a location's operational hours differ from its normal business hours. A special hour period must represent a range of less than 24 hours. The open_time and start_date must predate the close_time and end_date. The close_time and end_date can extend to 11:59 a.m. on the day after the specified start_date. For example, the following inputs are valid:

SpecialHours
Represents a set of time periods when a location's operational hours differ from its normal business hours.

StartUploadMediaItemDataRequest
Request message for Media.StartUploadMediaItemData.

TargetLocation
Represents a target location for a pending invitation.

TimeDimension
The dimension for which data is divided over.

TimeInterval
An interval of time, inclusive. It must contain all fields to be valid.

TimeOfDay
Represents a time of day. The date and time zone are either not significant or are specified elsewhere. An API may choose to allow leap seconds. Related types are google.type.Date and google.protobuf.Timestamp.

TimePeriod
Represents a span of time that the business is open, starting on the specified open day/time and closing on the specified close day/time. The closing time must occur after the opening time, for example later in the same day, or on a subsequent day.

TimeRange
A range of time. Data will be pulled over the range as a half-open inverval (that is, [start_time, end_time)).

TopDirectionSources
Top regions where driving-direction requests originated from.

TransferLocationRequest
Request message for Locations.TransferLocation.

UpsertAnswerRequest
Request message for QuestionsAndAnswers.UpsertAnswer

UrlAttributeValue
Values for an attribute with a value_type of URL.

Verification
A verification represents a verification attempt on a location.

VerificationOption
The verification option represents how to verify the location (indicated by verification method) and where the verification will be sent to (indicated by display data).

VerificationToken
Token generated by a vetted partner.

VerificationTokenGenerateCall
Generates a token for the provided location data as a vetted partner.

VerificationTokenMethods
A builder providing access to all methods supported on verificationToken resources. It is not used directly, but through the MyBusiness hub.

VerifyLocationRequest
Request message for Verifications.VerifyLocation.

VerifyLocationResponse
Response message for Verifications.VerifyLocation.

VettedPartnerInput
Input for VETTED_PARTNER verification.