Struct google_mybusiness4::api::Location

pub struct Location {
    pub ad_words_location_extensions: Option<AdWordsLocationExtensions>,
    pub additional_categories: Option<Vec<Category>>,
    pub additional_phones: Option<Vec<String>>,
    pub address: Option<PostalAddress>,
    pub attributes: Option<Vec<Attribute>>,
    pub labels: Option<Vec<String>>,
    pub language_code: Option<String>,
    pub latlng: Option<LatLng>,
    pub location_key: Option<LocationKey>,
    pub location_name: Option<String>,
    pub location_state: Option<LocationState>,
    pub metadata: Option<Metadata>,
    pub name: Option<String>,
    pub open_info: Option<OpenInfo>,
    pub price_lists: Option<Vec<PriceList>>,
    pub primary_category: Option<Category>,
    pub primary_phone: Option<String>,
    pub profile: Option<Profile>,
    pub regular_hours: Option<BusinessHours>,
    pub relationship_data: Option<RelationshipData>,
    pub service_area: Option<ServiceAreaBusiness>,
    pub special_hours: Option<SpecialHours>,
    pub store_code: Option<String>,
    pub website_url: Option<String>,
}

A location. See the [help center article] (https://support.google.com/business/answer/3038177) for a detailed description of these fields, or the category endpoint for a list of valid business categories.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations create accounts (request|response)

locations get accounts (response)

locations patch accounts (request|response)

locations transfer accounts (response)

Fields
ad_words_location_extensions: Option<AdWordsLocationExtensions>

Additional information that is surfaced in AdWords.

additional_categories: Option<Vec<Category>>

Additional categories to describe your business. Categories help your customers find accurate, specific results for services they're interested in. To keep your business information accurate and live, make sure that you use as few categories as possible to describe your overall core business. Choose categories that are as specific as possible, but representative of your main business.

additional_phones: Option<Vec<String>>

Up to two phone numbers (mobile or landline, no fax) at which your business can be called, in addition to your primary phone number.

address: Option<PostalAddress>

A precise, accurate address to describe your business location. PO boxes or mailboxes located at remote locations are not acceptable. At this time, you can specify a maximum of five address_lines values in the address.

attributes: Option<Vec<Attribute>>

Attributes for this location.

labels: Option<Vec<String>>

A collection of free-form strings to allow you to tag your business. These labels are NOT user facing; only you can see them. Limited to 255 characters (per label).

language_code: Option<String>

The language of the location. Set during creation and not updateable.

latlng: Option<LatLng>

User-provided latitude and longitude. When creating a location, this field is ignored if the provided address geocodes successfully. This field is only returned on get requests if the user-provided latlng value was accepted during create, or the latlng value was updated through the Google My Business website. This field cannot be updated.

location_key: Option<LocationKey>

A collection of keys that link this business to other Google properties, such as a Google+ page and Google Maps Places.

location_name: Option<String>

Location name should reflect your business's real-world name, as used consistently on your storefront, website, and stationery, and as known to customers. Any additional information, when relevant, can be included in other fields of the resource (for example, Address, Categories). Don't add unnecessary information to your name (for example, prefer "Google" over "Google Inc. - Mountain View Corporate Headquarters"). Don't include marketing taglines, store codes, special characters, hours or closed/open status, phone numbers, website URLs, service/product information, location/address or directions, or containment information (for example, "Chase ATM in Duane Reade").

location_state: Option<LocationState>

Output only. A set of booleans that reflects the state of a location.

metadata: Option<Metadata>

Output Only. Additional non-user-editable information.

name: Option<String>

Google identifier for this location in the form: accounts/{account_id}/locations/{location_id}

In the context of matches, this field will not be populated.

open_info: Option<OpenInfo>

A flag that indicates whether the location is currently open for business.

price_lists: Option<Vec<PriceList>>

Price list information for this location.

primary_category: Option<Category>

Category that best describes the core business this location engages in.

primary_phone: Option<String>

A phone number that connects to your individual business location as directly as possible. Use a local phone number instead of a central, call center helpline number whenever possible.

profile: Option<Profile>

Describes your business in your own voice and shares with users the unique story of your business and offerings.

regular_hours: Option<BusinessHours>

Operating hours for the business.

relationship_data: Option<RelationshipData>

All locations and chain related to this one.

service_area: Option<ServiceAreaBusiness>

Service area businesses provide their service at the customer's location. If this business is a service area business, this field describes the area(s) serviced by the business.

special_hours: Option<SpecialHours>

Special hours for the business. This typically includes holiday hours, and other times outside of regular operating hours. These override regular business hours.

store_code: Option<String>

External identifier for this location, which must be unique inside a given account. This is a means of associating the location with your own records.

website_url: Option<String>

A URL for this business. If possible, use a URL that represents this individual business location instead of a generic website/URL that represents all locations, or the brand.